requirements.txt
flask>=3.0.0
google-generativeai>=0.8.0
gTTS>=2.5.0
python-dotenv>=1.0.0
Pillow>=10.0.0
markdown>=3.5.0
